pyproject.toml
[project]
name = "linkforge-workspace"
version = "1.3.0" # x-release-please-version
description = "Universal Robotics Interoperability Workspace"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[tool.uv.workspace]
members = ["core", "platforms/blender"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
    "myst-parser>=2.0.0",
    "sphinx>=7.2.6",
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinxcontrib-mermaid>=0.9.2",
    "sphinx-design>=0.5.0",
    "fake-bpy-module-4.2",
]

# Shared Tool Configs
[tool.pytest.ini_options]
testpaths = ["tests/unit/core"]
pythonpath = ["core/src", "platforms/blender"]
addopts = ["--verbose", "--cov=linkforge_core", "--cov=platforms/blender/linkforge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM"]
ignore = ["E501", "B008", "C901", "B009", "N801"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

# Blender (bpy) stubs use Python 3.12 syntax.
# We set the parser to 3.12 above and skip deep inspection here.
[[tool.mypy.overrides]]
module = "bpy.*"
ignore_errors = true
ignore_missing_imports = true

# Enable full type checking for all LinkForge Blender modules
[[tool.mypy.overrides]]
module = "linkforge.blender.*"
ignore_errors = false

# Tiered Quality Control: Strict enforcement for production core, optimized for developer velocity in test suites.
[[tool.mypy.overrides]]
module = "tests.*"
check_untyped_defs = false
disallow_untyped_defs = false
ignore_missing_imports = true
ignore_errors = true

[dependency-groups]
dev = []
